Disable Selfrando and wrapper script for Firefox _helper_
Selfrando (used since 7.0a4) has a massive 12 seconds startup delay.
Disable it for the Firefox helper (which is only started to pass a URL
to the master, then it exits).

start-tor-browser hardcodes the SELFRANDO_write_layout_file environment
variable, so let's have the helper command start the firefox binary
directly. This also gets rid of a .1 second delay caused by the wrapper
script.