Don't let web pages see key-up events for our hotkeys
When a hotkey is pressed, Firefox insists on sending a key-up event to
the web page. This can lead to buggy behavior on some login forms (if
Ctrl-Shift-Enter is treated like Enter), and it's bad for privacy: Web
pages should not be notified when the user presses Ctrl-d to create a
bookmark. Even stock Firefox / Tor Browser (i.e. without Split Browser)
behaves this way.

So at least for our own hotkeys, use lower-level SDK APIs to handle both
key-down and key-up, and suppress the completed key combination's key-up
event as well. If the hotkey includes Ctrl, it will seem as if the user
pressed Ctrl and then released it without completing any combination.
Alt and Shift are invisible.