Make 'make install-persist' not break the DisposableVM side
Running 'make install-persist' in the DisposableVM template broke Split
Browser, due to a name conflict between /run/split-browser/cmd/ (managed
by systemd-tmpfiles and intended for use by the persistent side) and the
/run/split-browser/cmd socket (through which URLs are passed from the
qrexec service to the Firefox extension on the disposable side).

Rename the latter file to /run/split-browser/ext so that both parts can
safely be installed at the same time in the DisposableVM template.