Graphical qvm-move-to-vm on Ctrl-Shift-s
Instead of the CLI qvm-move-to-vm, call qvm-move-to-vm.kde (which will
fall back to qvm-move-to-vm.gnome). It should allow VM selection on both
R3.2 and R4.0.

Start it with a fresh login environment to avoid LD_LIBRARY_PATH related
incompatibilities in TB.

Pass * even when there are no files, so that the user will see an error
dialog rather than no response at all.