split-browser-bookmark: Fix XML-escaping with Bash 5.2+
Bash 5.2 added a patsub_replacement shopt (and enabled it by default...)
changing the meaning of an unquoted '&' - it is no longer a literal in a
pattern substitution's replacement string. This broke XML-escaping in
notification bodies:

<lt;
>gt;
"quot;
'apos;

Quote the ampersand. Actually overquote every part of pattern
substitution, just in case.