Use rofi instead of dmenu (temporarily?)
Even as a normal non-OverrideRedirect window, dmenu was still triggering
buggy behavior in the Qubes OS GUI protocol like repeating keystrokes,
e.g. a spurious Enter unintentionally confirming a login. This seems to
be more likely on on faster hardware. It's somehow related to a normal
window calling XGrabKeyboard() close to being focused by the window
manager.

Switch to rofi until this is sorted out.